:art: feat: integrate cache directly into repositories
