+++ src/infrastructure/cache/index.ts
@@ -1,6 +1,4 @@
 export { CacheService } from './CacheService';
 export { CacheKeys, CacheTTL } from './CacheKeys';
 export { CacheManager } from './CacheManager';
-export { CachedTransactionRepository } from './CachedTransactionRepository';
-export { CachedDashboardRepository } from './CachedDashboardRepository';
