+++ src/infrastructure/data/repositories/DashboardRepository.ts
@@ -1,6 +1,8 @@
 import { IDashboardRepository } from '../../../domain/repositories/IDashboardRepository';
 import { DashboardData, ChartData } from '../../../domain/entities/Dashboard';
 import { firebaseDashboardService } from '../../../../services/firebaseDashboardService';
+import { CacheService } from '../../cache/CacheService';
+import { CacheKeys, CacheTTL } from '../../cache/CacheKeys';
 
 export class DashboardRepository implements IDashboardRepository {
   async getDashboardData(
@@ -8,7 +10,21 @@ export class DashboardRepository implements IDashboardRepository {
     period: 'week' | 'month' | 'year',
     selectedMonth: string
   ): Promise<DashboardData> {
-    return firebaseDashboardService.getDashboardData(userId, period, selectedMonth);
+    const cacheKey = CacheKeys.DASHBOARD_DATA(userId, period, selectedMonth);
+    
+    // Tentar recuperar do cache
+    const cached = await CacheService.get<DashboardData>(cacheKey);
+    if (cached) {
+      return cached;
+    }
+
+    // Se não estiver em cache, buscar do serviço
+    const data = await firebaseDashboardService.getDashboardData(userId, period, selectedMonth);
+    
+    // Salvar no cache (TTL de 5 minutos)
+    await CacheService.set(cacheKey, data, CacheTTL.MEDIUM);
+    
+    return data;
   }
 
   async getChartData(
@@ -16,7 +32,21 @@ export class DashboardRepository implements IDashboardRepository {
     period: 'week' | 'month' | 'year',
     selectedMonth: string
   ): Promise<ChartData[]> {
-    return firebaseDashboardService.getChartData(userId, period, selectedMonth);
+    const cacheKey = CacheKeys.DASHBOARD_CHARTS(userId, period, selectedMonth);
+    
+    // Tentar recuperar do cache
+    const cached = await CacheService.get<ChartData[]>(cacheKey);
+    if (cached) {
+      return cached;
+    }
+
+    // Se não estiver em cache, buscar do serviço
+    const data = await firebaseDashboardService.getChartData(userId, period, selectedMonth);
+    
+    // Salvar no cache (TTL de 5 minutos)
+    await CacheService.set(cacheKey, data, CacheTTL.MEDIUM);
+    
+    return data;
   }
 }
