+++ src/infrastructure/di/container.ts
@@ -1,14 +1,14 @@
+import { TransactionRepository } from '../data/repositories/TransactionRepository';
 import { AuthRepository } from '../data/repositories/AuthRepository';
-import { CachedTransactionRepository } from '../cache/CachedTransactionRepository';
-import { CachedDashboardRepository } from '../cache/CachedDashboardRepository';
+import { DashboardRepository } from '../data/repositories/DashboardRepository';
 import { TransactionUseCases } from '../../domain/usecases/TransactionUseCases';
 import { AuthUseCases } from '../../domain/usecases/AuthUseCases';
 import { DashboardUseCases } from '../../domain/usecases/DashboardUseCases';
 
-// Repositórios com cache
-const transactionRepository = new CachedTransactionRepository();
-const authRepository = new AuthRepository(); // Auth não precisa de cache
-const dashboardRepository = new CachedDashboardRepository();
+// Repositórios (agora com cache integrado)
+const transactionRepository = new TransactionRepository();
+const authRepository = new AuthRepository();
+const dashboardRepository = new DashboardRepository();
 
 // Use Cases
 export const transactionUseCases = new TransactionUseCases(transactionRepository);
